Move birthday reader to open_city_profile package
Which is under the utils

Refs TILA-1983
